add auth check, user-post relation

+++ controllers/feed.js
@@ -2,6 +2,7 @@ const { validationResult } = require("express-validator");
 const Post = require("../models/post");
 const fs = require("fs");
 const path = require("path");
+const User = require("../models/user");
 function removeFile(path) {
   fs.unlink(path, (err) => {
     if (err) throw err;
@@ -35,8 +36,8 @@ exports.getPosts = async (req, res, next) => {
   }
 };
 exports.getPostDetails = async (req, res, next) => {
-  const { postId } = req.params;
   try {
+    const { postId } = req.params;
     const post = await Post.findById(postId);
     if (!post) throwNotFoundError("Post not found");
     return res.status(200).json({
@@ -49,26 +50,34 @@ exports.getPostDetails = async (req, res, next) => {
 
 //POST
 exports.postPosts = async (req, res, next) => {
-  const validatedResultOfRequest = validationResult(req);
-  const { title, content, author } = req.body;
-  const imageUrl = req.file?.path;
   try {
-    if (!imageUrl) return res.status(403).json({ error: "No image uploaded" });
+    const validatedResultOfRequest = validationResult(req);
+    const { title, content } = req.body;
+    const authorId = req.userId;
+    const imageUrl = req.file?.path;
+    if (!imageUrl) throw Error("No image uploaded");
     if (!validatedResultOfRequest.isEmpty()) {
       removeFile(getImageFullPath(req.file.filename));
-      return res.status(403).json({ errors: validatedResultOfRequest.array() });
+      throw validatedResultOfRequest.array();
     }
     const post = new Post({
       title: title,
       content: content,
-      author: author,
+      author: authorId,
       imageUrl: imageUrl,
     });
-
     const savedPost = await post.save();
     if (!savedPost) throw new Error("post not saved, try again later.");
+    const author = await User.findById(authorId);
+    if (!author) throw new Error("post saved but not to user.");
+    author.posts.push(savedPost._id);
+    const updatedUser = await author.save();
+    if (!updatedUser) {
+      await Post.findByIdAndDelete(savedPost._id);
+      throw Error("Could't save post,");
+    }
     res.status(201).json({
-      post,
+      post: savedPost,
     });
   } catch (error) {
     next(error);
@@ -82,6 +91,11 @@ exports.editPost = async (req, res, next) => {
   const image = req.file;
   try {
     const foundPost = await Post.findById(postId);
+    if (!foundPost) throw new Error("Post Not Found.");
+    if (foundPost.author.toString() !== req.userId.toString()) {
+      image && removeFile(image.path);
+      throw new Error("Not Authorized to update post.");
+    }
     const imageUrl = image?.path || foundPost.imageUrl;
     const updatedPost = await Post.findByIdAndUpdate(
       postId,
@@ -102,8 +116,18 @@ exports.editPost = async (req, res, next) => {
 exports.deletePost = async (req, res, next) => {
   const { postId } = req.params;
   try {
+    const [user, post] = await Promise.all([
+      User.findById(req.userId),
+      Post.findById(postId),
+    ]);
+    if (!post) throwNotFoundError("Post not found");
+    if (!user || user._id.toString() !== post.author.toString())
+      throw new Error("Not Authorized to delete post.");
     const deletedPost = await Post.findByIdAndDelete(postId);
-    if (!deletedPost) throwNotFoundError("Post not found");
+    user.posts.pull(postId);
+    const savedUser = await user.save();
+    if (!savedUser || !deletedPost) throw new Error("Couldn't delete post");
+    //add post again if user not saved
     removeFile(deletedPost.imageUrl);
     return res.status(200).json({ deletedPost });
   } catch (error) {
